[FIX] fix issues with the binary output format for the CCSDS121

+++ src/2.1-data_compression/include/NoCompression.h
@@ -2,6 +2,7 @@
 #define NC_H
 
 #include "Config.h"
+#include "BitOutputUtils.h"
 
 
 //#define NC_DEBUG 
@@ -27,6 +28,6 @@ printf("\n")
 #endif
 
 
-struct FCompressedData NoCompression(unsigned long int* Samples);
+struct FCompressedData NoCompression(unsigned int* Samples);
 
 #endif
